Generalize extracting feature
