+++ FCGR/setup.sh
@@ -243,20 +243,20 @@ cat << "EOF" > $table
 Type their corresponding number as single block of numbers (eg. 12), then press [ENTER]:
 
     1. CDS : coding sequences
-    2. LCR : low complexity regions (including transposable elements)
+    2. STR : Short Tandem Repeats
 EOF
 choice=fixed
 good_inputs=12
 abbrev="config/temp/abbrev.txt"
 cat << "EOF" > $abbrev
-NA CDS LCR
+NA CDS STR
 EOF
 out_dir="config/features/"
 nice="config/temp/nice.txt"
 cat << "EOF" > $nice
 NA
 Coding sequences (CDS)
-Low complexity regions (LCR)
+Short Tandem Repeats (STR)
 EOF
 
 confirm
