+++ scripts/download_genomes.sh
@@ -6,9 +6,12 @@
 
 # Will download whole concatenated genomes
 
-species=$( $1 | tr -d '{}')
-output_genome=$( $2 | tr -d '{}')
-output_table=$( $3 | tr -d '{}')
+species=$( $1 )
+output_genome=$( $2 )
+output_table=$( $3 )
+if [ $species != e_coli ]; then
+    output_repeats=$( $4 )
+fi
 
 declare -A all_accessions
 all_accessions=(["h_sapiens"]="ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/001/405/GCF_000001405.36_GRCh38.p10/"
@@ -22,8 +25,12 @@ all_accessions=(["h_sapiens"]="ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/000/00
 accession=${all_accessions[$species]}
 
 IFS='/' read -r -a array <<< $accession
-whole_genome=$(echo ${array[9]}'_genomic.fna.gz' )
-feature_table=$(echo ${array[9]}'_feature_table.txt.gz' )
+whole_genome=$( echo ${array[9]}'_genomic.fna.gz' )
+feature_table=$( echo ${array[9]}'_feature_table.txt.gz' )
+
+if [ $species != e_coli ]; then
+    repeats=$( echo ${array[9]}'_rm.out.gz' )
+fi
 
 wget --no-use-server-timestamps $accession$whole_genome
 gunzip < $whole_genome > $output_genome
@@ -32,3 +39,9 @@ rm $whole_genome
 wget --no-use-server-timestamps $accession$feature_table
 gunzip < $feature_table > $output_table
 rm $feature_table
+
+if [ $species != e_coli ]; then
+    wget --no-use-server-timestamps $accession$repeats
+    gunzip < $repeats > $output_repeats
+    rm $repeats
+fi
